+++ tests/unit/test_main.py
@@ -10,6 +10,7 @@ from napps.kytos.flow_manager.exceptions import (
 )
 from napps.kytos.of_core.v0x04.flow import Flow as Flow04
 from pyof.v0x04.asynchronous.error_msg import ErrorType
+from pyof.v0x04.common.header import Type
 from pyof.v0x04.controller2switch.flow_mod import FlowModCommand
 
 from kytos.core.helpers import now
@@ -984,11 +985,13 @@ class TestMain(TestCase):
 
         flow = MagicMock()
         flow.as_dict.return_value = {}
+        flow.header.message_type = Type.OFPT_FLOW_MOD
         flow.xid = 1
         self.napp._flow_mods_sent[flow.xid] = (flow, "add")
 
         mock_ev = MagicMock()
-        mock_ev.event.content = {"destination": switch}
+        mock_ev.message = flow
+        mock_ev.content["destination"] = switch
         min_wait = 0.2
         multiplier = 2
         assert self.napp._retry_on_openflow_connection_error(
@@ -1012,6 +1015,7 @@ class TestMain(TestCase):
 
         flow = MagicMock()
         flow.as_dict.return_value = {}
+        flow.header.message_type = Type.OFPT_FLOW_MOD
         flow.xid = 1
         self.napp._flow_mods_sent[flow.xid] = (flow, "add")
 
@@ -1019,7 +1023,7 @@ class TestMain(TestCase):
         self.napp._flow_mods_retry_count[flow.xid] = (3, now(), 10)
 
         mock_ev = MagicMock()
-        mock_ev.event.content = {"destination": switch}
+        mock_ev.message = flow
         min_wait = 0.2
         assert not self.napp._retry_on_openflow_connection_error(
             mock_ev,
@@ -1030,24 +1034,33 @@ class TestMain(TestCase):
         )
         assert mock_send.call_count == 1
 
-    def test_retry_on_openflow_connection_error_early_return(self):
+    def test_retry_on_openflow_connection_error_early_return_max_retries(self):
         """Test retry on openflow connection error early returns."""
         max_retries = 0
         min_wait = 0.2
         multiplier = 2
+        mock, flow = MagicMock(), MagicMock()
+        flow.header.message_type = Type.OFPT_FLOW_MOD
+        mock.message = flow
         with self.assertRaises(ValueError) as exc:
             self.napp._retry_on_openflow_connection_error(
-                {}, max_retries, min_wait, multiplier
+                mock, max_retries, min_wait, multiplier
             )
         assert "should be > 0" in str(exc.exception)
 
         self.napp._flow_mods_sent = {}
-        mock = MagicMock()
+        mock, flow = MagicMock(), MagicMock()
+        flow.header.message_type = Type.OFPT_FLOW_MOD
+        mock.message = flow
         with self.assertRaises(ValueError) as exc:
             self.napp._retry_on_openflow_connection_error(
                 mock, max_retries + 1, min_wait, multiplier
             )
         assert "not found on flow mods sent" in str(exc.exception)
+
+    def test_retry_on_openflow_connection_error_early_return_msg_type(self):
+        """Test retry on openflow connection error early returns."""
+        assert not self.napp._retry_on_openflow_connection_error(MagicMock())
 
     @patch("napps.kytos.flow_manager.main.Main._send_napp_event")
     def test_send_openflow_connection_error(self, mock_send):